로직, UI 분리 완료
