+++ App.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import RootNavigation from './src/Navigation/RootNavigation';
+import FirebaseInit from './src/Api/GoogleLogin/FirebaseInit';
+
+const App = () => {
+  <FirebaseInit />;
+  return <RootNavigation />;
+};
+
+export default App;
